grade avec model
